Migrate to null safety. Add example project
